# flowcatalyst-typescript/apps/platform-frontend/openapi/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
...
    ClientConfigListResponse:
      description: List of client configurations
      type: object
      properties:
        clientConfigs:
          type: array
          items:
            $ref: '#/components/schemas/ClientConfigResponse'
          description: Client configurations
        total:
          type: integer
          format: int32
          description: Total count
    ClientConfigRequest:
      type: object
      properties:
        enabled:
          type: boolean
        baseUrlOverride:
          type: string
        websiteOverride:
          type: string
        config:
          type: object
          additionalProperties: {}
    ClientConfigResponse:
      type: object
      description: Application client configuration
      properties:
        id:
          type: string
          description: Config ID
        applicationId:
          type: string
          description: Application ID
        clientId:
          type: string
          description: Client ID
        clientName:
          type: string
          description: Client name
        clientIdentifier:
          type: string
          description: Client identifier
        enabled:
          type: boolean
          description: Whether enabled for this client
        baseUrlOverride:
          type: string
          description: Base URL override for this client
        websiteOverride:
          type: string
          description: Website override for this client
        effectiveBaseUrl:
          type: string
          description: Effective base URL (override or default)
        effectiveWebsite:
          type: string
          description: Effective website (override or default)
        config:
          type: object
          additionalProperties: {}
          description: Additional configuration
    ClientDto:
      description: Client details
      type: object
      properties:
        id:
          type: string
          description: Client ID
          examples:
            - cli_0ABC123DEF456
        name:
          type: string
          description: Client name
          examples:
            - Acme Corporation
        identifier:
          type: string
          description: Unique client identifier/slug
          examples:
            - acme-corp
        status:
          description: Client status
          type: string
          $ref: '#/components/schemas/ClientStatus'
        statusReason:
          type: string
          description: Reason for current status
        statusChangedAt:
          description: When status was last changed
          type: string
          $ref: '#/components/schemas/Instant'
        createdAt:
          description: Creation timestamp
          type: string
          $ref: '#/components/schemas/Instant'
        updatedAt:
          description: Last update timestamp
          type: string
          $ref: '#/components/schemas/Instant'
...
    ClientListResponse:
      description: Client list response
      type: object
      properties:
        clients:
          type: array
          items:
            $ref: '#/components/schemas/ClientDto'
          description: List of clients
        total:
          type: integer
          format: int32
          description: Total count
    ClientListResponse1:
      description: List of clients
      type: object
      properties:
        clients:
          type: array
          items:
            $ref: '#/components/schemas/ClientDto1'
          description: Client list
        total:
          type: integer
          format: int32
          description: Total count
...
    ClientListResponse3:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ClientResponse'
    ClientResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        identifier:
          type: string
        status:
          type: string
    ClientStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
        - SUSPENDED
...
    CreateClientRequest:
      description: Request to create a new client
      type: object
      required:
        - name
        - identifier
      properties:
        name:
          type: string
          description: Client display name
          examples:
            - Acme Corporation
          pattern: \S
          maxLength: 255
        identifier:
          type: string
          description: Unique client identifier/slug
          examples:
            - acme-corp
          pattern: \S
          minLength: 2
          maxLength: 100
    CreateClientRequest1:
      description: Create client request
      type: object
      required:
        - name
        - identifier
      properties:
        name:
          type: string
          description: Client name
        identifier:
          type: string
          description: Unique identifier/slug
...
    ErrorResponse:
      description: Error response for client errors (400)
      type: object
      properties:
        code:
          type: string
          description: Error code for programmatic handling
          examples:
            - VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error message
          examples:
            - Invalid input provided
        details:
          type: object
          additionalProperties: {}
          description: Additional error details
...
    ForbiddenResponse:
      description: Forbidden error response (403)
      type: object
      properties:
        code:
          type: string
          description: Error code
          examples:
            - FORBIDDEN
        message:
          type: string
          description: Human-readable error message
          examples:
            - Insufficient permissions
        requiredPermission:
          type: string
          description: Required permission if applicable
          examples:
            - application:delete
...
    MyClientsResponse:
      description: List of accessible clients
      type: object
      properties:
        clients:
          type: array
          items:
            $ref: '#/components/schemas/MyClientDto'
          description: Clients the user can access
        total:
          type: integer
          format: int32
          description: Total count
    NotFoundResponse:
      description: Not found error response (404)
      type: object
      properties:
        code:
          type: string
          description: Error code
          examples:
            - NOT_FOUND
        message:
          type: string
          description: Human-readable error message
          examples:
            - Application not found
        resourceType:
          type: string
          description: Type of resource that was not found
          examples:
            - application
        resourceId:
          type: string
          description: ID that was searched for
          examples:
            - app_0ABC123DEF456
...
    UnauthorizedResponse:
      description: Unauthorized error response (401)
      type: object
      properties:
        code:
          type: string
          description: Error code
          examples:
            - UNAUTHORIZED
        message:
          type: string
          description: Human-readable error message
          examples:
            - Authentication required
...
    UpdateClientRequest:
      description: Request to update a client
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Client display name
          examples:
            - Acme Corporation
          pattern: \S
          maxLength: 255
...
  securitySchemes:
    SecurityScheme:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication
paths:
...
  /api/admin/applications/{id}/clients:
    get:
      summary: Get client configurations for an application
      operationId: getApplicationClientConfigs
      tags:
        - BFF - Application Admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Client configurations retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientConfigListResponse'
        '404':
          description: Application not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
  /api/admin/applications/{id}/clients/{clientId}:
    put:
      summary: Configure application for a specific client
      operationId: configureApplicationForClient
      tags:
        - BFF - Application Admin
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientConfigRequest'
        required: true
      responses:
        '200':
          description: Client configuration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientConfigResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Application or client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
...
  /api/admin/clients:
    get:
      summary: List all clients
      description: Returns all clients regardless of status
      operationId: listClients
      tags:
        - BFF - Client Admin
      parameters:
        - description: Filter by status
          name: status
          in: query
          schema:
            $ref: '#/components/schemas/ClientStatus'
      responses:
        '200':
          description: List of clients
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientListResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
    post:
      summary: Create a new client
      operationId: createClient
      tags:
        - BFF - Client Admin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClientRequest'
        required: true
      responses:
        '201':
          description: Client created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDto'
        '400':
          description: Invalid request or identifier already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
...
  /api/admin/clients/{id}:
    put:
      summary: Update client details
      operationId: updateClient
      tags:
        - BFF - Client Admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClientRequest'
        required: true
      responses:
        '200':
          description: Client updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDto'
        '400':
          description: Invalid request or client ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
    get:
      summary: Get client by ID
      operationId: getClient
      tags:
        - BFF - Client Admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Client details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDto'
        '400':
          description: Invalid client ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
...
  /api/clients:
    get:
      summary: Get accessible clients
      description: "Returns the list of clients the authenticated principal has access\
        \ to, based on their scope."
      tags:
        - Clients
      responses:
        '200':
          description: List of accessible clients
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientListResponse3'
        '401':
          description: Not authenticated
  /api/clients/{id}:
    get:
      summary: Get client by ID
      description: Returns a specific client if the caller has access to it.
      tags:
        - Clients
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Client found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
        '400':
          description: Invalid client ID format
        '401':
          description: Not authenticated
        '403':
          description: Access denied
        '404':
          description: Client not found
...
  /api/me/clients:
    get:
      summary: Get my accessible clients
      description: Returns clients the authenticated user can access based on their
        scope
      operationId: getMyClients
      tags:
        - User
      parameters:
        - name: fc_session
          in: cookie
          schema:
            type: string
      responses:
        '200':
          description: List of accessible clients
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyClientsResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
  /api/me/clients/{clientId}:
    get:
      summary: Get a specific accessible client
      description: Returns client details if the user has access to it
      operationId: getMyClient
      tags:
        - User
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
        - name: fc_session
          in: cookie
          schema:
            type: string
      responses:
        '200':
          description: Client details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyClientDto'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedResponse'
        '403':
          description: No access to this client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
        '404':
          description: Client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundResponse'
...
  /api/sdk/clients:
    get:
      summary: List clients
      operationId: sdkListClients
      tags:
        - SDK - Clients
      parameters:
        - description: Filter by status
          name: status
          in: query
          schema:
            $ref: '#/components/schemas/ClientStatus'
      responses:
        '200':
          description: List of clients
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientListResponse1'
        '401':
          description: Not authenticated
        '403':
          description: Insufficient permissions
    post:
      summary: Create a new client
      operationId: sdkCreateClient
      tags:
        - SDK - Clients
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClientRequest1'
        required: true
      responses:
        '201':
          description: Client created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDto1'
        '400':
          description: Invalid request
        '401':
          description: Not authenticated
        '403':
          description: Insufficient permissions
  /api/sdk/clients/{id}:
    put:
      summary: Update client details
      operationId: sdkUpdateClient
      tags:
        - SDK - Clients
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClientRequest1'
        required: true
      responses:
        '200':
          description: Client updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDto1'
        '404':
          description: Client not found
        '401':
          description: Not authenticated
        '403':
          description: Insufficient permissions
        '400':
          description: Bad Request
    get:
      summary: Get client by ID
      operationId: sdkGetClient
      tags:
        - SDK - Clients
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Client details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDto1'
        '404':
          description: Client not found
        '401':
          description: Not authenticated
        '403':
          description: Insufficient permissions
...
  /api/sdk/principals/{id}/clients:
    get:
      summary: Get principal's granted clients
      operationId: sdkGetPrincipalClients
      tags:
        - SDK - Principals
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of client access grants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientAccessListResponse'
        '404':
          description: Principal not found
        '401':
          description: Not authenticated
        '403':
          description: Insufficient permissions
  /api/sdk/principals/{id}/clients/{clientId}:
    delete:
      summary: Revoke client access from principal
      operationId: sdkRevokeClientAccess
      tags:
        - SDK - Principals
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Access revoked
        '404':
          description: Grant not found
        '401':
          description: Not authenticated
        '403':
          description: Insufficient permissions
    post:
      summary: Grant client access to principal
      operationId: sdkGrantClientAccess
      tags:
        - SDK - Principals
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Access granted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientAccessGrantDto'
        '400':
          description: Grant already exists
        '404':
          description: Principal or client not found
        '401':
          description: Not authenticated
        '403':
          description: Insufficient permissions
...
  /auth/client/accessible:
    get:
      summary: List accessible clients for current user
      tags:
        - Client Selection
      parameters:
        - name: fc_session
          in: cookie
          schema:
            type: string
      responses:
        '200':
          description: List of accessible clients
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessibleClientsResponse'
        '401':
          description: Not authenticated
  /auth/client/current:
    get:
      summary: Get current client context
      tags:
        - Client Selection
      parameters:
        - name: fc_session
          in: cookie
          schema:
            type: string
      responses:
        '200':
          description: Current client info
          content:
            application/json:
              schema: {}
        '401':
          description: Not authenticated
  /auth/client/switch:
    post:
      summary: Switch to a different client context
      tags:
        - Client Selection
      parameters:
        - name: fc_session
          in: cookie
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwitchClientRequest'
        required: true
      responses:
        '200':
          description: New token issued with client context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwitchClientResponse'
        '401':
          description: Not authenticated
        '403':
          description: Access denied to client
        '400':
          description: Bad Request
...
info:
  title: FlowCatalyst Platform API
  version: 1.0.0-SNAPSHOT
  description: FlowCatalyst Platform REST API
servers:
  - url: http://localhost:8080
    description: Auto generated value
  - url: http://0.0.0.0:8080
    description: Auto generated value